Fix test-case failures due to character reference parsing bugs.
More completely, change the entire strucutre of the NCR array to be a multi-dimensional array with each dimension having as a key one character and a value possible further steps, and sometimes 'codepoint', which that point represents. Hence, to get &amp; now, you would need $ncrs['a']['m']['p'][';']['codepoint']. 'codepoint' is required so we can cope with both &amp and &amp; (which means it can't just be a value of &amp).

This also removes Data::getNamedCharacterReferenceMaxLength(), as it is now useless.